fix(docs): Invalid schema link
seems like PortableContacts.net has changed owners and no longer maintained by Joseph, according to [Archive.org][archive] the latest draft was still published on Apr 2016.

I've [contacted][tweet] Joseph to clarify, but in the meantime, its best not to link to that domain and reference the IETF link instead.

[archive]: http://web.archive.org/web/20160426223008/http://portablecontacts.net/draft-spec.html
[tweet]: https://twitter.com/AhmadNassri/status/817334438203891712